버그신고, 택시팟 생성, 참가, 비밀번호 변경 기능 구현.
To do : 택시팟 참가 시 리스트를 보여주는 기능, 히스토리 저장 자동화

+++ src/main/java/kr/hs/dsm/java/taxipot_backend/repository/TaxipotRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface TaxipotRepository extends CrudRepository<TaxiPot, Long> {
+public interface TaxipotRepository extends CrudRepository<TaxiPot, Integer> {
     public List<TaxiPot> findAll();
 }
